Move ci_to_gubernator main pkg into sub directory
