+++ toolbox/ci2gubernator/BUILD.bazel
@@ -1,0 +1,13 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["ci2gubernator.go"],
+    importpath = "istio.io/test-infra/toolbox/ci2gubernator",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//sisyphus:go_default_library",
+        "//toolbox/util:go_default_library",
+        "@com_github_marcacohen_gcslock//:go_default_library",
+    ],
+)
