+++ toolbox/ci2gubernator/main/BUILD.bazel
@@ -1,0 +1,19 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["main.go"],
+    importpath = "istio.io/test-infra/toolbox/ci2gubernator/main",
+    visibility = ["//visibility:private"],
+    deps = [
+        "//toolbox/ci2gubernator:go_default_library",
+        "//toolbox/util:go_default_library",
+    ],
+)
+
+go_binary(
+    name = "ci2gubernator",
+    embed = [":go_default_library"],
+    importpath = "istio.io/test-infra/toolbox/ci2gubernator/main",
+    visibility = ["//visibility:public"],
+)
