--- toolbox/ci_to_gubernator/BUILD.bazel
@@ -1,24 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["ci_to_gubernator.go"],
-    importpath = "istio.io/test-infra/toolbox/ci_to_gubernator",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//sisyphus:go_default_library",
-        "//toolbox/util:go_default_library",
-        "@com_github_marcacohen_gcslock//:go_default_library",
-    ],
-)
-
-go_library(
-    name = "main",
-    srcs = ["main.go"],
-    importpath = "istio.io/test-infra/toolbox/ci_to_gubernator/main",
-    visibility = ["//visibility:private"],
-    deps = [
-        "//toolbox/ci_to_gubernator:go_default_library",
-        "//toolbox/util:go_default_library",
-    ],
-)
